Remove util.py since it only had one function used in one place
