+++ api/src/multiaccess/memberbooth.py
@@ -4,10 +4,20 @@ from membership.membership import get_membership_summary
 from service.db import db_session
 from service.error import NotFound
 from membership.models import Member, Key
-from .util import member_to_response_object
 
 
 logger = getLogger("makeradmin")
+
+
+def member_to_response_object(member: Member):
+    return {
+        'member_id': member.member_id,
+        'member_number': member.member_number,
+        'firstname': member.firstname,
+        'lastname': member.lastname,
+        'end_date': max((span.enddate for span in member.spans)).isoformat() if len(member.spans) > 0 else None,
+        'keys': [{'key_id': key.key_id, 'rfid_tag': key.tagid} for key in member.keys],
+    }
 
 
 def memberbooth_response_object(member: Member, membership_data):
