--- api/src/multiaccess/util.py
@@ -1,13 +1,0 @@
-from membership.models import Member
-
-
-def member_to_response_object(member: Member):
-    return {
-        'member_id': member.member_id,
-        'member_number': member.member_number,
-        'firstname': member.firstname,
-        'lastname': member.lastname,
-        'end_date': max((span.enddate for span in member.spans)).isoformat() if len(member.spans) > 0 else None,
-        'keys': [{'key_id': key.key_id, 'rfid_tag': key.tagid} for key in member.keys],
-    }
-
